code/deployment/api/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow>=8.0.0
aiofiles

# ML & Deep Learning
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0

# Image Processing
numpy>=1.19.0
opencv-python-headless>=4.5.0

# Development & Quality
pytest==7.4.3
black==23.11.0
flake8==6.1.0
